Fix printf arguments to match format strings and vice-versa (PR#18578).
git-svn-id: https://svn.r-project.org/R/trunk@85535 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/inspect.c
@@ -135,7 +135,7 @@ static void inspect_tree(int pre, SEXP v, int deep, int pvec) {
     switch (TYPEOF(v)) {
     case VECSXP: case STRSXP: case LGLSXP: case INTSXP: case RAWSXP:
     case REALSXP: case CPLXSXP: case EXPRSXP:
-	Rprintf("(len=%ld, tl=%ld)", XLENGTH(v), XTRUELENGTH(v));
+	Rprintf("(len=%ld, tl=%ld)", (long)XLENGTH(v), (long)XTRUELENGTH(v));
     }
     if (TYPEOF(v) == ENVSXP) /* NOTE: this is not a trivial OP since it involves looking up things
 				in the environment, so for a low-level debugging we may want to
